Add files via upload
The location of waste disposal sites in Colombia is available in a shapefile format, which can be modified for use in environmental, logistical, territorial planning, and health analyses, among others. 

This data was created based on the information available from the Corporaciones Autónomas Regionales  (CAR) and the Sistema Único de información de Servicios Públicos (SUI). 

The maps were developed using available geographic information and corroborated through satellite imagery. The sites are classified as Landfills, Cells, and Dumps in accordance with municipal agreements.